+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/entity/animal/AnimalFeedOrderOnly.java
@@ -12,6 +12,7 @@ import org.hibernate.annotations.NotFound;
 import org.hibernate.annotations.NotFoundAction;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -201,7 +202,7 @@ public class AnimalFeedOrderOnly extends BaseEntity {
      * 单价
      */
     @Column(name = "unit_fee")
-    private Integer unitFee;
+    private BigDecimal unitFee;
 
     /**
      * 入库备注
